+++ backend/canisters/community/CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Bot's can't access private channels they don't own ([#8181](https://github.com/open-chat-labs/open-chat/pull/8181))
 
+### Removed
+
+- Remove `gate` which is superseded by `gate_config` ([#8186](https://github.com/open-chat-labs/open-chat/pull/8186))
+
 ## [[2.0.1793](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1793-community)] - 2025-06-13
 
 ### Added